refactor: change capitalization of the scoreGauge component file
