+++ frontend/src/lib/components/scoreGauge/ScoreGauge.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import type {ScoreGaugeProps} from './scoreGauge.type';
+  import type {ScoreGaugeProps} from './ScoreGauge.type';
   import {getUUID} from '$lib/utils/components';
   import {_} from 'svelte-i18n';
