+++ frontend/src/lib/components/scoreGauge/ScoreGauge.type.ts
